Extracted ExclusionsBuilder into separate class
