+++ src/main/java/io/github/mfvanek/pg/connection/PrimaryHostDeterminerImpl.java
@@ -31,8 +31,8 @@ public class PrimaryHostDeterminerImpl implements PrimaryHostDeterminer {
      */
     @Override
     public boolean isPrimary(@Nonnull final PgConnection pgConnection) {
-        LOGGER.debug("Executing query: {}", SQL_QUERY);
         Objects.requireNonNull(pgConnection, "pgConnection");
+        LOGGER.debug("Executing on host {} query: {}", pgConnection.getHost(), SQL_QUERY);
         if (pgConnection.getHost().cannotBePrimary()) {
             return false;
         }
